server/sam-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==2.1.3
Pillow==11.1.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
